New license, new project name.
The license is now MIT since I like it better for this kind of project.

The name is now 'salt' since I like it better and think that if it's not
more descriptive, it's at least less overloaded/misleading than 'NaCl'
(OK, but seriously, both djb NaCl *and* google NaCl are acronyms, for
totally unrelated things, but the pronunciation for both of them is
apparently 'salt.' Whatever.) Maybe this will confuse less people. Or not.

I'm not changing the module structure since I don't think it's necessary.

The name change really just reflects the name of the Haskell package,
not the underlying bindings. If anything, the structure should stay the
same, since right now, all the NaCl* modules are literally bindings to
the NaCl library and nothing else, and that's at least straightforward.

I'm open to disagreement on these points but I think it's not that big
a deal since nobody depends on this yet.