A different approach to building the NaCl source
This moves us away from the bootstrap.sh script. The basic overview is
that now, Cabal invokes a custom './configure' script that builds nacl,
copies it to an output directory, and outputs a salt.buildinfo with the
necessary paths.

This makes the installation much more smooth, and it does
auto-detection of a build existing. Unfortunately it also means I have
to add the entire NaCl source code to the source distro, but maybe it
would be reasonable to untar it?