Redo the whole nonce API
This version of the API basically uses a type class to propagate the
type of a nonce. Nobody should need to use ugly nonce* utilities to get
lengths anymore, and you still can't confuse two kinds of nonces, but
still have a shared API among the 3 different kinds. This is actually a
nice bit of code reduction as well (no internal module, etc.)

This adds a dependency on the 'tagged' package, since I prefer it to
passing 'undefined' around to get the size of some particular nonce.

As a result of the API change, one of the tests was no longer valid.

Needs documentation updates; benchmarks and examples are also currently
broken.