API Change: parametrize keys
The 'Key' type is now parametrized by two (phantom) type variables: an
's' parameter referring to secrecy, and an 'i' parameter referring to
index. Together, these distinguish keys as to whether they are
public/private, and what operations they're used for (signing,
authentication, etc.)

This is more orthogonal and also actually simpler in my eyes. It is also
more descriptive.

The tests are currently a complete mess and need to be rewritten in
complete honesty. See issue #24.

Some documentation is probably broken from this.

Closes #22
Closes #17
Closes #16

Signed-off-by: Austin Seipp <mad.one@gmail.com>